Stock pushed

+++ src/main/java/tn/test/spring/Repository/StockRepository.java
@@ -3,5 +3,9 @@ package tn.test.spring.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.test.spring.Entity.Stock;
 
-public interface StockRepository extends JpaRepository<Stock, Long> {
+public interface StockRepository extends JpaRepository<Stock, Integer> {
+
+
+
 }
+
